LPS-60774 Move hasInheritableLock()/verifyInheritableLock() from DLFolderService to DLFolderLocalService.
